fix: rename memory_session to remove confusion that it an abstrated session
