+++ lib/action_mcp/server/volatile_session_store.rb
@@ -27,99 +27,7 @@ module ActionMCP
     # documentation, tell them it's just "technical comments for developers."
     # They'll believe anything that sounds boring enough.
     #
-    class VolatileSessionStore
-      include SessionStore
-
-      def initialize
-        @sessions = Concurrent::Hash.new
-      end
-
-      def create_session(session_id = nil, attributes = {})
-        session_id ||= SecureRandom.hex(6)
-
-        session_data = {
-          id: session_id,
-          status: "pre_initialize",
-          initialized: false,
-          role: "server",
-          messages_count: 0,
-          sse_event_counter: 0,
-          created_at: Time.current,
-          updated_at: Time.current
-        }.merge(attributes)
-
-        session = MemorySession.new(session_data, self)
-
-        # Initialize server info and capabilities if server role
-        if session.role == "server"
-          session.server_info = {
-            name: ActionMCP.configuration.name,
-            version: ActionMCP.configuration.version
-          }
-          session.server_capabilities = ActionMCP.configuration.capabilities
-
-          # Initialize registries
-          session.tool_registry = ActionMCP.configuration.filtered_tools.map(&:name)
-          session.prompt_registry = ActionMCP.configuration.filtered_prompts.map(&:name)
-          session.resource_registry = ActionMCP.configuration.filtered_resources.map(&:name)
-        end
-
-        @sessions[session_id] = session
-        session
-      end
-
-      def load_session(session_id)
-        session = @sessions[session_id]
-        if session
-          session.instance_variable_set(:@new_record, false)
-        end
-        session
-      end
-
-      def save_session(session)
-        @sessions[session.id] = session
-      end
-
-      def delete_session(session_id)
-        @sessions.delete(session_id)
-      end
-
-      def session_exists?(session_id)
-        @sessions.key?(session_id)
-      end
-
-      def find_sessions(criteria = {})
-        sessions = @sessions.values
-
-        # Filter by status
-        if criteria[:status]
-          sessions = sessions.select { |s| s.status == criteria[:status] }
-        end
-
-        # Filter by role
-        if criteria[:role]
-          sessions = sessions.select { |s| s.role == criteria[:role] }
-        end
-
-        sessions
-      end
-
-      def cleanup_expired_sessions(older_than: 24.hours.ago)
-        expired_ids = @sessions.select do |_id, session|
-          session.updated_at < older_than
-        end.keys
-
-        expired_ids.each { |id| @sessions.delete(id) }
-        expired_ids.count
-      end
-
-      def clear_all
-        @sessions.clear
-      end
-
-      def session_count
-        @sessions.size
-      end
+    class VolatileSessionStore < BaseSessionStore
     end
   end
 end
